generate code sample using 127.0.0.1 due to Java/.NET weirdness
Java & .NET SDK's result in an error when we try to use the host param
(e.g., some IP address passed in as the -host command-line param), so
for now all code samples will be generated using "127.0.0.1" as the
host whether or not there's a session or not.

This might be revisited once we start using DNS names like
cb-01.couchbase.live instead of an IP address for the -host param.